Revert "Split out `_SingleBuild` as `Build` and `BuildImpl` as `BuildSeries`.…"
This reverts commit 224dd81c5ba9001d6731a9871849ec68fdd491f1.
