+++ build_runner/test/server/serve_handler_test.dart
@@ -15,7 +15,6 @@ import 'package:build_runner/src/server/server.dart';
 import 'package:build_runner_core/build_runner_core.dart';
 import 'package:build_runner_core/src/asset_graph/graph.dart';
 import 'package:build_runner_core/src/asset_graph/node.dart';
-import 'package:build_runner_core/src/generate/build_phases.dart';
 import 'package:build_runner_core/src/generate/options.dart';
 import 'package:build_runner_core/src/generate/performance_tracker.dart';
 import 'package:build_runner_core/src/package_graph/target_graph.dart';
@@ -108,7 +107,7 @@ void main() {
     final packageGraph = buildPackageGraph({rootPackage('a'): []});
     readerWriter = TestReaderWriter(rootPackage: packageGraph.root.name);
     assetGraph = await AssetGraph.build(
-      BuildPhases([]),
+      [],
       <AssetId>{},
       <AssetId>{},
       packageGraph,
@@ -123,7 +122,7 @@ void main() {
             packageGraph,
             defaultRootPackageSources: defaultRootPackageSources,
           ),
-          BuildPhases([]),
+          [],
           'a',
         ),
       ),
