+++ build_runner_core/lib/src/asset_graph/optional_output_tracker.dart
@@ -5,7 +5,7 @@
 import 'package:build/build.dart';
 
 import '../../build_runner_core.dart';
-import '../generate/build_phases.dart';
+import '../generate/phase.dart';
 import '../package_graph/target_graph.dart';
 import '../util/build_dirs.dart';
 import 'graph.dart';
@@ -35,7 +35,7 @@ class OptionalOutputTracker {
   final TargetGraph _targetGraph;
   final Set<String> _buildDirs;
   final Set<BuildFilter> _buildFilters;
-  final BuildPhases _buildPhases;
+  final List<BuildPhase> _buildPhases;
 
   OptionalOutputTracker(
     this._assetGraph,
