+++ build_runner_core/lib/src/package_graph/apply_builders.dart
@@ -12,7 +12,6 @@ import 'package:build_config/build_config.dart';
 import 'package:graphs/graphs.dart';
 import 'package:logging/logging.dart';
 
-import '../generate/build_phases.dart';
 import '../generate/exceptions.dart';
 import '../generate/phase.dart';
 import '../validation/config_validation.dart';
@@ -293,7 +292,7 @@ final _logger = Logger('ApplyBuilders');
 /// Builders may be filtered, for instance to run only on package which have a
 /// dependency on some other package by choosing the appropriate
 /// [BuilderApplication].
-Future<BuildPhases> createBuildPhases(
+Future<List<BuildPhase>> createBuildPhases(
   TargetGraph targetGraph,
   Iterable<BuilderApplication> builderApplications,
   Map<String, Map<String, dynamic>> builderConfigOverrides,
@@ -371,7 +370,7 @@ Future<BuildPhases> createBuildPhases(
     );
   }
 
-  return BuildPhases([...inBuildPhases, ...collapsedPostBuildPhase]);
+  return <BuildPhase>[...inBuildPhases, ...collapsedPostBuildPhase];
 }
 
 Iterable<BuildPhase> _createBuildPhasesWithinCycle(
