+++ build_runner_core/test/asset_graph/graph_test.dart
@@ -11,7 +11,6 @@ import 'package:build/build.dart';
 import 'package:build_config/build_config.dart';
 import 'package:build_runner_core/src/asset_graph/graph.dart';
 import 'package:build_runner_core/src/asset_graph/node.dart';
-import 'package:build_runner_core/src/generate/build_phases.dart';
 import 'package:build_runner_core/src/generate/phase.dart';
 import 'package:crypto/crypto.dart';
 import 'package:test/test.dart';
@@ -49,7 +48,7 @@ void main() {
     group('simple graph', () {
       setUp(() async {
         graph = await AssetGraph.build(
-          BuildPhases([]),
+          [],
           <AssetId>{},
           <AssetId>{},
           fooPackageGraph,
@@ -213,7 +212,7 @@ void main() {
 
     group('with buildPhases', () {
       var targetSources = const InputSet(exclude: ['excluded.txt']);
-      final buildPhases = BuildPhases([
+      final buildPhases = [
         InBuildPhase(
           TestBuilder(buildExtensions: appendExtension('.copy', from: '.txt')),
           'foo',
@@ -228,7 +227,7 @@ void main() {
             generateFor: const InputSet(),
           ),
         ]),
-      ]);
+      ];
       final primaryInputId = makeAssetId('foo|file.txt');
       final excludedInputId = makeAssetId('foo|excluded.txt');
       final primaryOutputId = makeAssetId('foo|file.txt.copy');
@@ -578,7 +577,7 @@ void main() {
     test('overlapping build phases cause an error', () async {
       expect(
         () => AssetGraph.build(
-          BuildPhases(List.filled(2, InBuildPhase(TestBuilder(), 'foo'))),
+          List.filled(2, InBuildPhase(TestBuilder(), 'foo')),
           {makeAssetId('foo|file')},
           <AssetId>{},
           fooPackageGraph,
@@ -605,7 +604,7 @@ void main() {
           digestReader.testing.writeString(id, 'contents of $id');
         }
         final graph = await AssetGraph.build(
-          BuildPhases([
+          [
             InBuildPhase(
               TestBuilder(buildExtensions: replaceExtension('.txt', '.a.txt')),
               'foo',
@@ -623,7 +622,7 @@ void main() {
               'foo',
               hideOutput: false,
             ),
-          ]),
+          ],
           sources,
           <AssetId>{},
           fooPackageGraph,
@@ -658,7 +657,7 @@ void main() {
           digestReader.testing.writeString(id, 'contents of $id');
         }
         final graph = await AssetGraph.build(
-          BuildPhases([
+          [
             InBuildPhase(
               TestBuilder(buildExtensions: appendExtension('.1', from: '.txt')),
               'foo',
@@ -668,7 +667,7 @@ void main() {
               'foo',
               targetSources: const InputSet(include: ['lib/*.txt']),
             ),
-          ]),
+          ],
           sources,
           <AssetId>{},
           fooPackageGraph,
@@ -688,7 +687,7 @@ void main() {
         final nodeToRead = AssetId('foo', 'lib/a.1');
         final outputReadingNode = AssetId('foo', 'lib/b.2');
         final lastPrimaryOutputNode = AssetId('foo', 'lib/b.3');
-        final buildPhases = BuildPhases([
+        final buildPhases = [
           InBuildPhase(
             TestBuilder(buildExtensions: replaceExtension('.txt', '.1')),
             'foo',
@@ -701,7 +700,7 @@ void main() {
             TestBuilder(buildExtensions: replaceExtension('.2', '.3')),
             'foo',
           ),
-        ]);
+        ];
         final sources = {makeAssetId('foo|lib/b.anchor')};
         for (final id in sources) {
           digestReader.testing.writeString(id, 'contents of $id');
@@ -751,7 +750,7 @@ void main() {
         final generatedDart = AssetId('a', 'lib/a.g.dart');
         final generatedPart = AssetId('a', 'lib/a.g.part');
         final toBeGeneratedDart = AssetId('a', 'lib/A.g.dart');
-        final buildPhases = BuildPhases([
+        final buildPhases = [
           InBuildPhase(
             TestBuilder(buildExtensions: replaceExtension('.dart', '.g.part')),
             'a',
@@ -762,7 +761,7 @@ void main() {
             ),
             'a',
           ),
-        ]);
+        ];
         final packageGraph = buildPackageGraph({rootPackage('a'): []});
         final graph = await AssetGraph.build(
           buildPhases,
