+++ build_runner_core/test/environment/create_merged_dir_test.dart
@@ -13,7 +13,6 @@ import 'package:build_runner_core/src/asset_graph/graph.dart';
 import 'package:build_runner_core/src/asset_graph/node.dart';
 import 'package:build_runner_core/src/asset_graph/optional_output_tracker.dart';
 import 'package:build_runner_core/src/environment/create_merged_dir.dart';
-import 'package:build_runner_core/src/generate/build_phases.dart';
 import 'package:build_runner_core/src/generate/options.dart';
 import 'package:build_runner_core/src/generate/phase.dart';
 import 'package:build_runner_core/src/package_graph/target_graph.dart';
@@ -23,7 +22,7 @@ import 'package:test/test.dart';
 void main() {
   group('createMergedDir', () {
     late AssetGraph graph;
-    final phases = BuildPhases([
+    final phases = [
       InBuildPhase(
         TestBuilder(buildExtensions: appendExtension('.copy', from: '.txt')),
         'a',
@@ -32,7 +31,7 @@ void main() {
         TestBuilder(buildExtensions: appendExtension('.copy', from: '.txt')),
         'b',
       ),
-    ]);
+    ];
     final sources = {
       makeAssetId('a|lib/a.txt'): 'a',
       makeAssetId('a|web/b.txt'): 'b',
