+++ build_runner_core/test/generate/asset_tracker_test.dart
@@ -8,8 +8,7 @@ import 'dart:io';
 import 'package:build/build.dart';
 import 'package:build_runner_core/build_runner_core.dart';
 import 'package:build_runner_core/src/asset_graph/graph.dart';
-import 'package:build_runner_core/src/generate/asset_tracker.dart';
-import 'package:build_runner_core/src/generate/build_phases.dart';
+import 'package:build_runner_core/src/generate/build_definition.dart';
 import 'package:build_runner_core/src/package_graph/target_graph.dart';
 import 'package:package_config/package_config.dart';
 import 'package:path/path.dart' as p;
@@ -44,7 +43,7 @@ void main() {
       var reader = ReaderWriter(packageGraph);
       var aId = AssetId('a', 'web/a.txt');
       assetGraph = await AssetGraph.build(
-        BuildPhases([]),
+        [],
         {aId},
         <AssetId>{},
         packageGraph,
@@ -64,7 +63,7 @@ void main() {
       assetTracker = AssetTracker(reader, targetGraph);
       var updates = await assetTracker.collectChanges(assetGraph);
       await assetGraph.updateAndInvalidate(
-        BuildPhases([]),
+        [],
         updates,
         'a',
         (_) async {},
