+++ build_runner_core/test/package_graph/apply_builders_test.dart
@@ -6,7 +6,6 @@ import 'dart:async';
 import 'package:_test_common/common.dart';
 import 'package:build/build.dart';
 import 'package:build_config/build_config.dart';
-import 'package:build_runner_core/src/generate/build_phases.dart';
 import 'package:build_runner_core/src/generate/exceptions.dart';
 import 'package:build_runner_core/src/generate/phase.dart';
 import 'package:build_runner_core/src/package_graph/apply_builders.dart';
@@ -30,7 +29,7 @@ void main() {
       var phases = await createBuildPhases(targetGraph, builderApplications, {
         'b:cool_builder': {'option_a': 'a', 'option_c': 'c'},
       }, false);
-      for (final phase in phases.phases.cast<InBuildPhase>()) {
+      for (final phase in phases.cast<InBuildPhase>()) {
         expect((phase.builder as CoolBuilder).optionA, equals('a'));
         expect((phase.builder as CoolBuilder).optionB, equals('defaultB'));
         expect((phase.builder as CoolBuilder).optionC, equals('c'));
@@ -79,7 +78,7 @@ void main() {
               },
               true,
             );
-            for (final phase in phases.phases.cast<InBuildPhase>()) {
+            for (final phase in phases.cast<InBuildPhase>()) {
               expect(
                 (phase.builder as CoolBuilder).optionA,
                 equals('global a'),
@@ -119,7 +118,7 @@ void main() {
         false,
       );
       expect(phases, hasLength(1));
-      expect((phases.phases.first as InBuildPhase).package, 'a');
+      expect((phases.first as InBuildPhase).package, 'a');
     });
 
     test('honors appliesBuilders', () async {
@@ -277,7 +276,7 @@ void main() {
     });
 
     group('autoApplyBuilders', () {
-      Future<BuildPhases> createPhases({
+      Future<List<BuildPhase>> createPhases({
         Map<String, TargetBuilderConfig>? builderConfigs,
       }) async {
         var packageGraph = buildPackageGraph({
@@ -333,7 +332,7 @@ void main() {
         );
         expect(phases, hasLength(1));
         expect(
-          phases.phases.first,
+          phases.first,
           isA<InBuildPhase>()
               .having((p) => p.package, 'package', 'a')
               .having(
